+++ src/Http/Requests/CreateTaxon.php
@@ -18,13 +18,14 @@ use Vanilo\Admin\Contracts\Requests\CreateTaxon as CreateTaxonContract;
 
 class CreateTaxon extends FormRequest implements CreateTaxonContract
 {
-    public function rules()
+    public function rules(): array
     {
         return [
             'name' => 'required|string|min:2|max:255',
             'slug' => 'sometimes|nullable|string|max:255',
             'parent_id' => 'nullable|exists:taxons,id',
             'priority' => 'nullable|integer',
+            'is_active' => 'nullable|boolean',
             'ext_title' => 'sometimes|nullable|string|max:511',
             'meta_description' => 'sometimes|nullable|string|max:4096',
             'meta_keywords' => 'sometimes|nullable|string|max:1024',
@@ -38,7 +39,7 @@ class CreateTaxon extends FormRequest implements CreateTaxonContract
         ];
     }
 
-    public function authorize()
+    public function authorize(): bool
     {
         return true;
     }